docs: Update Readme (#71)
