+++ align_anything/datasets/text_image_to_text_image/preference.py
