--- align_anything/datasets/ti_to_ti/__init__.py
@@ -1,20 +1,0 @@
-# Copyright 2024 PKU-Alignment Team. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-"""Dataset classes for text to text training."""
-
-
-from align_anything.datasets.ti_to_ti.supervised import *
-from align_anything.datasets.ti_to_ti.preference import *
-from align_anything.datasets.ti_to_ti.prompt_only import *
