+++ align_anything/trainers/text_image_to_text_image/__init__.py
