+++ align_anything/trainers/text_image_to_text_image/ppo.py
@@ -27,7 +27,7 @@ import torch
 from transformers import GenerationConfig
 from transformers.integrations.deepspeed import HfDeepSpeedConfig
 
-from align_anything.datasets.ti_to_ti import PromptOnlyTokenizedDataset, SupervisedTokenizedDataset
+from align_anything.datasets.text_image_to_text_image import PromptOnlyTokenizedDataset, SupervisedTokenizedDataset
 from align_anything.models.pretrained_model import load_pretrained_models
 from align_anything.models.pretrained_model_with_value import load_pretrained_model_with_value_head
 from align_anything.trainers.text_to_text.ppo import PPOTrainer as PPOTextTrainer
@@ -167,7 +167,7 @@ def main():
     torch.cuda.set_device(current_device)
 
     # read default configs from the yaml file
-    task = os.path.join('ti_to_ti', 'ppo')
+    task = os.path.join('text_image_to_text_image', 'ppo')
     dict_cfgs, ds_cfgs = read_cfgs(mode='train', task=task)
 
     # get custom configs from command line
