+++ align_anything/trainers/text_image_to_text_image/sft.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""Trainer for DPO training."""
+"""Trainer for supervised training."""
 
 
 import argparse
@@ -23,52 +23,40 @@ import deepspeed
 import torch
 from transformers.integrations.deepspeed import HfDeepSpeedConfig
 
-from transformers import AutoModelForCausalLM
-from align_anything.datasets.text_to_text.preference import PreferenceBatch, PreferenceDataset
-from align_anything.datasets.ti_to_ti.preference import PreferenceTokenizedDataset
+from align_anything.datasets.text_image_to_text_image import SupervisedTokenizedDataset
 from align_anything.models.pretrained_model import load_pretrained_models
-from align_anything.trainers.text_to_text.dpo import DPOTrainer as DPOtextTrainer
+from align_anything.trainers.text_to_text.sft import SupervisedTrainer as SupervisedtextTrainer
 from align_anything.utils.multi_process import get_current_device
 from align_anything.utils.tools import (
     custom_cfgs_to_dict,
     dict_to_namedtuple,
-    gather_log_probabilities,
     read_cfgs,
     seed_everything,
     update_dict,
 )
 
 
-class DPOTrainer(DPOtextTrainer):
+class SuperviseTrainer(SupervisedtextTrainer):
 
     def init_datasets(self) -> None:
         """Initialize training and evaluation datasets."""
         self.train_dataloader, self.eval_dataloader = self.get_dataloaders(
-            PreferenceTokenizedDataset, PreferenceTokenizedDataset
+            SupervisedTokenizedDataset, SupervisedTokenizedDataset
         )
 
     def init_models(self) -> None:
         """Initialize model and tokenizer."""
-        if self.ds_train_cfgs['zero_optimization']['stage'] == 3:
-            self.dstchf_train = HfDeepSpeedConfig(self.ds_train_cfgs)
-        if self.ds_eval_cfgs['zero_optimization']['stage'] == 3:
-            self.dsechf_eval = HfDeepSpeedConfig(self.ds_eval_cfgs)
+        if self.ds_train_cfgs is not None and self.ds_train_cfgs['zero_optimization']['stage'] == 3:
+            self.dstchf = HfDeepSpeedConfig(self.ds_train_cfgs)
         self.model, self.tokenizer, self.processor = load_pretrained_models(
             self.cfgs.model_cfgs.model_name_or_path,
             model_max_length=self.cfgs.model_cfgs.model_max_length,
-            padding_side='left',
-            trust_remote_code=self.cfgs.train_cfgs.trust_remote_code,
+            padding_side='right',
+            trust_remote_code=True,
             freeze_mm_proj=self.cfgs.train_cfgs.freeze_mm_proj,
             freeze_vision_tower=self.cfgs.train_cfgs.freeze_vision_tower,
             freeze_language_model=self.cfgs.train_cfgs.freeze_language_model,
         )
-        self.reference_model, _, _ = load_pretrained_models(
-            self.cfgs.model_cfgs.model_name_or_path,
-            model_max_length=self.cfgs.model_cfgs.model_max_length,
-            padding_side='left',
-            trust_remote_code=self.cfgs.train_cfgs.trust_remote_code,
-        )
-
 
 
 def main():
@@ -78,7 +66,7 @@ def main():
     torch.cuda.set_device(current_device)
 
     # read default configs from the yaml file
-    task = os.path.join('ti_to_ti', 'dpo')
+    task = os.path.join('text_image_to_text_image', 'sft')
     dict_cfgs, ds_cfgs = read_cfgs(mode='train', task=task)
 
     # get custom configs from command line
@@ -95,7 +83,7 @@ def main():
     seed_everything(cfgs.train_cfgs.seed)
 
     # finetune the model
-    trainer = DPOTrainer(cfgs=cfgs, ds_cfgs=ds_cfgs)
+    trainer = SuperviseTrainer(cfgs=cfgs, ds_cfgs=ds_cfgs)
     trainer.train()
     trainer.save()
