+++ assets/text_image_to_text_image/bathroom_generated.png
