+++ assets/text_image_to_text_image/pizza_generated.png
