+++ projects/any_to_text/README.md
@@ -108,6 +108,7 @@ deepspeed \
 ### Model Weights
 
 - [Image-Text Multimodal Language Model](https://huggingface.co/PKU-Alignment/llama3.1-8b-instruct-vision)
+- [Image-Audio-Text Multimodal Language Model](https://huggingface.co/PKU-Alignment/llama3.1-8b-vision-audio)
 
 ### Model Evaluation
