+++ projects/text_image_to_text_image/.gitignore
