+++ projects/text_image_to_text_image/pre_tokenize_parallel_example.py
@@ -46,6 +46,7 @@ from align_anything.utils.multi_process import get_current_device
 from align_anything.utils.template_registry import get_template_class
 from align_anything.utils.tools import right_padding
 from datasets import load_dataset
+import argparse
 
 ALLOWED_ATTRIBUTES = ['split_token']
 DEFAULT_SPLIT_TOKEN = 'ASSISTANT:'
@@ -71,7 +72,6 @@ def format_sample_cham(raw_sample: dict[str, Any]) -> dict[str, Any]:
     assistant_prompt: str = '\nASSISTANT:{output}'
     split_token: str = 'ASSISTANT:'
     separator: str = '###'
-    # print(raw_sample)
     input_text = raw_sample['question']
     output_text = raw_sample['response']
     input_img = raw_sample['image_url']
@@ -233,11 +233,18 @@ def process_data(gpu, input_data, model_path, output_path, cache_dir):
     print(f"GPU {gpu} processed {len(local_output_paths)} messages")
 
 def main():
-    input_path = "input.json" # change this to your input path
-    output_path = "output.pt"
+    parser = argparse.ArgumentParser()
+    parser.add_argument("--input_path", type=str, required=True)
+    parser.add_argument("--output_path", type=str, required=True)
+    parser.add_argument("--model_path", type=str, required=True)
+    parser.add_argument("--cache_dir", type=str, required=True)
     
-    model_path = "path_to_model"
-    cache_path = ".cache"
+    args = parser.parse_args()
+
+    input_path = args.input_path
+    output_path = args.output_path
+    model_path = args.model_path
+    cache_path = args.cache_dir
     
     # if cache dir does not exist, make one
     if not os.path.exists(cache_path):
@@ -277,6 +284,6 @@ def main():
     
     torch.save(all_data, output_path)
         
-# do main
+
 if __name__ == "__main__":
     main()
